make Bitcoin immutable and add intern pool
"Immutability" was implemented by setting slots and overriding
__setattr__.  This should discourage most attempts at mutation.

Interning is sensible for an immutable object and has the benefits
of saving memory and faster equality checks.

We did need to implement a deconstruct method for Django as Django's
`deconstructible` decorator overrides __new__, and we had to implement
our own __new__ for interning.